try: use swr for comments
